requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# 数据库
sqlalchemy==2.0.23

# 文档解析
python-docx==1.1.0
PyPDF2==3.0.1

# AI/ML (可选)
openai==1.3.5
langchain==0.0.340
sentence-transformers==2.2.2

# 工具库
pydantic==2.5.0
python-dotenv==1.0.0

# 其他
requests==2.31.0
